Refactor time in EE genesis #587

+++ programs/create-genesis-ee/event_engine_genesis.cpp
@@ -12,7 +12,7 @@ static abi_def create_messages_abi() {
         "vote_info", "", {
             {"voter", "name"},
             {"weight", "int16"},
-            {"time", "uint64"}
+            {"time", "uint32"}
         }
     });
 
@@ -21,7 +21,7 @@ static abi_def create_messages_abi() {
             {"account", "name"},
             {"title", "string"},
             {"body", "string"},
-            {"time", "uint64"}
+            {"time", "uint32"}
         }
     });
